Improving HTML to text
Adding a recursive function to delete all *unwanted tags* from the DOM
tree (instead of extracting the *wanted tags*, as was done in the
implementation up to this point).

Then, the entire resulting html soup is transfered to text with
markdown-elements, leading to crawler output as close as possible to the
actual rendered website text.