Fixing bugs in HTML decomposition, text extraction
**HTML tree decomposition**
There was a bug, where `for el in tag.children` was used instead of `for
el in list(tag.children)`.
When deleting/decomposing tags further down the loop, the iterator is not
reliable - tags will get skipped!

**Text extraction**
html2text has a bug, where some input (in the observed case some datetime
string) will be accumulated in the internal state of the library, and added
to the end of each output of h.handle().
This was mostly resolved by creating a new html2text handler each
iteration. Tests showed this is extremely fast and has no impact on the
speed of the crawler.

I tested *markdownify* as an alternative, but while markdownify appears more stable and produces less markdown errors, html2text offers a significantly better (i.e. closer to the website formatting) formatted output text.