Bugfixes/Stability: Browser reload, buffer, tag matching
* Browser reload:
Up to this point, a single browser instance was used for the whole
scraping process. Due to this, the crawler could be blocked using a
cookie: After a certain number of page visits, the website would return
400 message codes/not load anymore.
Now, each 100 visited pages, the browser is reinitialized, clearing all
cookies.
No idea if the browser can still be tracked across reloads theoretically,
but the 'rate limiting by cookie'-problem is solved now.
* Buffer:
The buffer was not written to file / flushed when reaching the end of
the scraping process, leading to discarding of valid texts. Now such a
flush() function exists and is called when exiting the crawling loop.
* tag matching:
Now, name:*name-value* allows for a *name-value* consisting of multiple
words. Only tags are matched whose *name* matches all of the words in
*name-value*.
---